Authentication + upload file

+++ delete.php
@@ -3,5 +3,10 @@
 // Suppression du véhicule
 require("header.php");
 
-$vehicleController->delete($_GET["id"]);
-echo "<script>window.location.href='index.php'</script>";
+if (!$_SESSION || !$_SESSION["email"]) {
+    echo "<script>window.location.href='index.php'</script>";
+} else {
+
+    $vehicleController->delete($_GET["id"]);
+    echo "<script>window.location.href='index.php'</script>";
+}
